requirements.txt
streamlit==1.32.0
openai==1.12.0
pyaudio==0.2.14
scipy==1.12.0
numpy==1.26.3
python-dotenv==1.0.0
requests==2.31.0
SpeechRecognition==3.10.1
pydub==0.25.1
httpx==0.24.1
pydantic==2.5.2
ormsgpack==1.3.0 